pyproject.toml
[project]
name = "weaviate-agents"
version = "0.4.3"
description = "The official sub-package for the Weaviate Agents project."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "rich>=13.9.4",
    "weaviate-client>=4.11.0",
]


[build-system]
requires = [
    "setuptools >= 65",
    "setuptools_scm[toml] >6.2",
    "wheel >= 0.38.1",
]
build-backend = 'setuptools.build_meta'

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pre-commit>=4.1.0",
    "pytest-cov>=6.0.0",
    "pytest>=8.3.4",
    "ruff>=0.9.6",
    "tomli>=2.2.1",
]

[tool.setuptools.packages.find]
where = ["."]  # Search in current directory
exclude = [
    "bash*",
    "tests*",
    "docs*",
    "scripts*",
]

[tool.setuptools]
license-files = []
